+++ BraintreeCard/BTCard.m
@@ -75,6 +75,10 @@
         p[@"expiration_year"] = self.expirationYear;
     }
 
+    if (self.cardholderName) {
+        p[@"cardholder_name"] = self.cardholderName;
+    }
+
     if (self.expirationMonth && self.expirationYear) {
         p[@"expiration_date"] = [NSString stringWithFormat:@"%@/%@", self.expirationMonth, self.expirationYear];
     }
@@ -265,6 +269,7 @@
      "      brand"
      "      expirationMonth"
      "      expirationYear"
+     "      cardholderName"
      "      last4"
      "      bin"
      "      binData {"