+++ BraintreeCore/Braintree-Version.h
@@ -1,1 +1,1 @@
-#define BRAINTREE_VERSION (@"4.34.0")
+#define BRAINTREE_VERSION (@"4.35.0")